Add first test
Adding a successful step for signing in and viewing the sessions page.

We will want to abstract away the sign in step but this proves we can use Wiremock in Playwright integration tests for now

We need to await all stubs to ensure the requist to the WireMock Admin API is resolved before proceeding with the test.