Populate date values in form
The date input is slightly fiddlier to manage in requests as the query is made
up of 3 separate parts. We also need to pass these parts back to the view as a
list of `items` in the format the nunjucks component expects them. We should
include any errors in these items as well, so we need to validate the
individual date parts (if all parts are required), alongside formatting.

I have created a `GovukFrontendDateInput` class that can handle some of this
logic and formatting for us. We can pass our query into the constructor, along
with the correct key, and could access formatted items. We could also store the
iso date value on here, and add validation.