Refactor GovUkFrontendDateInput and add date validations
The new date validation functions can be called independently and operate
solely on their given arguments. As a result, these can be implemented as
static methods on the GovukFrontendDateInput class and we can remove the need
for the class to hold state.

We also decouple the class from a query object and is just concerned with dates.

* Add static validation methods for completeness and validity
* Refactor items method to be static

These date utils have been copied from
https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-bail-ui