Update session search e2e test to include session page
This includes viewing a session in the scope of 'searching for a session'.

Have created a step to encapsulate some logic as with the session search.

Currently we are not seeding any data for these tests, so I am just finding the
first session link on the page, and asserting that the next page includes the
selected project title and some appointments.