Update project details data on check project details page
Now that the `AppointmentsDto` has been updated, we can display more data on the page.

Creating a new object to return to the view with only the data we need, and
with formatted date. This is slightly easier to write out in the view, and has
the benefit of not needing to be updated in the view (where there are no
typechecks) if the Dto ever changes.