Run accessibility tests when each page is loaded
We use cypress-axe to test each page as they are visited in our
integration tests. When an issue is found it will cause the test to
fail. We also add logging of accessibility issues to the console.

We include a rule which was inherited from the CAS services which allows us to
add back links outside of the `main` tag which is in line with govuk guidance.

Guidance:
> Always place back links at the top of a page, before the <main> element.
> Placing them here means that the ‘Skip to main content’ link allows the user
> to skip all navigation links, including the back link.