Add factories for offender and appointment
We may want to create different types of offenders in tests, but I am
envisioning that we will need a different factory for each.

If we need to change the offender type created in the appointment, we can pass
in a different object for the offender property. For example:

```
const appointment = appointmentFactory.build({ offender: offenderLimitedFactory.build() })
```