Update objects with new Dto properties
For now, I am not using a factory object in the Integration test mocks as we are hard asserting against some of these values in tests where we are expecting a different format.

We could improve this by creating a factory instance but just passing in values for those specific properties.