Add utility to convert a list of objects to the format needed for a GOV.UK Select Input.
Previously we have been mapping to this format inline, but this will reduce
changes if the component ever changed, and also allows us to add an unselected
value at the start of a list of options (Such as 'Choose supervisor').

As we have already introduced a class which maps to the correct format for a
GOV.UK component, following the same pattern here.