Update project details page to accept an expected appointment
When we are specifying response data we will need to assert against this in the
page object model.

Adding an optional constructor parameter to take an AppointmentDto.

Note: the 'shouldContainProjectDetails()' date will fail because I have written
out the expected date. Don't really like an application util to get the
expected value to be displayed.