Rename e2e test
This test will cover a full update appointment journey, so renaming
accordingly. At the moment we aren't creating data in the tests, so we are
stepping through from a search to find an appointment. We may be able to split
these steps out into two tests in the future if we seed data for the tests.