Introduce addSecondsToTime util
We currently store the start and end times in the following format HH:MM:SS. In
the interest of simplicity and matching Delius, we only want to present the
time as HH:MM.

As a result, when a time is given in the format HH:MM, we want to be able to
add seconds back to the time for consistency in our database.

This will be used in a follow-up piece of work when we update the appointment
endpoint.