Refactor test page models to accept appointment or session
We assert the title (usually name) on every page load. Previously we were
checking for a hard-coded value, now this will be taken from the object passed
in.

This will also be useful for testing a journey with a limited access offender
as we make use of the offender model which handles retrieving the name.