Introduce base appointment page model
We have developed a standalone page model to handle view and validation logic for each page in the appointment update form. This introduces a common base model to reduce some repetition across pages.

I plan to introduce a form url search parameter to help track the current in progress form across pages, so having a common place where the different page paths are assigned to the viewData will help simplify this.