Introduce query model for every page
This removes the need to cast query items from 'ParsedQs' types, and instead
define the properties we expect on the form body. Adding a base type with a
'form' property, which will be used to pass along the current in progress form
id via a hidden input or url search parameter.