Implement form save on each page
This finds the current in progress form (or creates a new one), adds form data
for that page to the form object and sends it to the forms API endpoint.

We need to use a regex for the Wiremock form stub, as the id is generated
internally in our application, so we have no way of knowing what it is.

We may want to stub specific requests in some requests (if for example
navigating to an existing form).