Add form property to view and route data
Once a form is in progress its id will be passed between pages as a URL search
param. This value will be passed to the submit handler as a hidden input in the
view.

A page model should always be instantiated with a query object, which will
either be the `req.query` or `req.body` depending if it is handling a GET or
POST action. The form property should be on both objects, but could be
undefined - for example if it is the first page of a form.