Create new in progress form on first page of form
Previously the form service created a blank form object if there was no existing formId when navigating to any page in the update appointment form journey. However this introduces some validation complications, so changing to explicitly create a new form in the controller methd of the first page in the journey.