Update e2e test to check for selected enforcement action
I expect we will add tests against other details with the implementation of CPB-124

Playwright doesn't have an obvious way to find values in descriptive lists, and
descriptive list items have no aria role - so using the html emmet as a
locator. The structure should match that of the GOV.UK component.