Make enforcement readonly
We will now only display guidance that an enforcement has been selected.

During testing we will still 'select' an enfocement from the list of
enforcement actions in the UI. I have moved this to be managed in submit of the
enforcement page. We will no longer capture any form data for enforcement, and
instead find the fixed one we want to save in the update by its code.

When creating test data, we will always need an enforcement action with this
code. For this we can use the `enforcementActionsFactory`.