Fix page verification for appointment page integration tests
These weren't fully testing we were on the right step of the appointment update form, and as a result we missed one test that wasn't behaving as expected.

Adding an empty function we can override at Page level, which will run at the
end of `checkOnPage`. This enables us to check for a h2 or form label on a page
where the h1 is the same on each page.