Store selected supervisor object in form
Following the pattern developed for storing contact outcome and enforcement
action, and storing the complete supervisor object on the in progress form,
rather than just the code.

This will enable us to easily display the name of the selected supervisor,
rather than performing an additional look up on the confirm page.