Display supervisor full name on confirm page instead of code
We want to switch to calculating this value from the `name` object on the
selected supervisor rather than the `fullName` property, but this works for
now.

Note: We have hardcoded the expected value selected in the e2e test, which may
result in failing tests when we change our data source. However, as the way in
which a supervisor is assigned may well change this feels fine for now.