Fix back to session link
From project details page

We were wrapping an existing link with a query with an extra query string
deliminator (`?`) because we were adding a formId to the link, which isn't
needed for this link, so overriding the value in the implementation class.

There is probably a better way of changing this just for the
project details page, but this should work to resolve the bug.