Fix `pathWithQuery` to return a valid URL
Here we fix `pathWithQuery` so that it always returns a valid URL and
never one with a trailing ? or indeed multiple ? characters.

We use `pathWithQuery` to construct the back link used across the app,
and the back link was in fact broken on the Check Project Details page.
The integration tests did not pick this up because the time formatting
function, `DateTimeFormats.isoDateToUIDate`, was able to function
with a *full ISO date* like `2025-12-09T07:14:30.845Z?` even with a
trailing ? character, but did *not* work with a shortened date like
`2025-12-09?` with a trailing ? character. Having updated the
`appointmentFactory` to produce shortened dates, the problem was then
able to be reproduced in the integration tests.