Create an appointment form object on rendering first page
The appointment form object will now contain appointment data as saved in
nDelius, to reduce logic about what to show on each page.

There is risk that the appointment could become out of date, but we expect
these forms to be short lived as there is no way to return to it, and we are
saving a version to flag an issue if the appointment has since been updated in
delius.