Add version to appointment form
As we are populating the whole form object on the first page of the form, there
is a chance that the delius version could change in between (which would have
also previously been the case). Instead capturing the version at the point of
creating the form, so we can through an error on submit if there is a
difference. The version submitted should be that on the form object, and not
the appointment.