Remove padding of date parts from date validation
We were previously hiding invalid dates without `0` padding as this method pads
parts and then tries to convert to a date. The problem is, we were only using
this format to 'validate' the date, but were not sending the same string to the
server, where the API would throw an error for dates without '0' padding.

I have added a step to update the query itself, before validating it, but we
also want to remove any formatting fro the validation step itself.